BaseModel and TimeModel (#525)
